Push basics are working!
Works exactly like Launchpad, only the repeat buttons are at the top.

TODO:
- Better use of colours [EASY]
- Use Push dedicated buttons for existing functions (play, rec, repeat,
  double etc.) without losing Launchpad-style mapping [MEDIUM]
- Display support (see notes below) [MEDIUM/HARD]
- Knob support [HARD?]

NOTES:
The hard work of talking to the display using SysEx was implemented
in  JS by Jurgen Mossgraber on his Bitwig Studio scripts.
We just need to port his implementation to something
that works for LoopDrop and hook the appropriate updater code.

https://github.com/git-moss/Push4Bitwig
https://github.com/git-moss/Push4Bitwig/blob/master/push/Display.js